Remove duplicate spaces in ban request messages
When a senior supervisor+ reacts approve/deny on a message in #ban-requests-queue, preemptively remove the duplicate whitespace using String.replaceAll("\\s+", " ") [searches for at least 2 spaces, replaces with 1 space) before splitting into the banRequestMessageContent array. This is to avoid empty values in the array, which can lead to errors when the final ban message is formatted/built and sent in the commands channel.