Update roles to use Discord IDs (long) instead of String
IDs are safer than hard-coded strings and will protect the code in case there are any role name changes in the future. The roles' data types were changed from String to long and some more helper functions were added to support this change.
This commit also changes two of the RDSS messages in the mod-alerts channel to grab the role mention by ID now that the hard-coded String doesn't exist.