Allow moderators to use quick-mute reactions on deleted or censored messages
If a 30m or 60m quick-mute reaction is used in the message-logs or censored-and-spam-logs channel by a moderator+, the original author of the message is temporarily muted according to which reaction was used, with the logged message's evidence as the reason in the infraction. Similar to how censored ban messages work currently. I copied from the ban-from-logs function to accomplish this, and added a `muteDuration` parameter that is passed in when a reaction is used.
Also, made some corrections in my previous comments that weren't 100% accurate or needed minor fixing.